move category parameters to framework package
