+++ packages/framework/src/Resources/translations/messages.cs.po
@@ -1663,6 +1663,12 @@ msgstr "Pro vytvoření produktu doplňte chybějící nastavení. Více informa
 msgid "Filter by"
 msgstr "Filtrovat podle"
 
+msgid "Filter parameters"
+msgstr "Parametry filtru"
+
+msgid "Filter parameters closed by default"
+msgstr "Defaultně zavřené parametry ve filtru"
+
 msgid "Finish the assigning"
 msgstr "Dokončit přiřazování"
 
@@ -2595,6 +2601,9 @@ msgstr "Parametry - vše"
 
 msgid "Parameters - view"
 msgstr "Parametry - zobrazení"
+
+msgid "Parameters order in category"
+msgstr "Pořadí parametrů v kategorii"
 
 msgid "Parameters overview"
 msgstr "Přehled parametrů"
