+++ project-base/app/src/Model/Category/CategoryData.php
@@ -8,19 +8,10 @@ use Shopsys\FrameworkBundle\Model\Category\CategoryData as BaseCategoryData;
 
 /**
  * @property \App\Model\Category\Category|null $parent
+ * @property \App\Model\Product\Parameter\Parameter[] $parametersCollapsed
  */
 class CategoryData extends BaseCategoryData
 {
-    /**
-     * @var \App\Model\Product\Parameter\Parameter[]
-     */
-    public $parametersCollapsed;
-
-    /**
-     * @var int[]|null[]
-     */
-    public $parametersPosition;
-
     /**
      * @var \App\Model\Category\Category[]
      */
@@ -30,7 +21,6 @@ class CategoryData extends BaseCategoryData
     {
         parent::__construct();
 
-        $this->parametersPosition = [];
         $this->linkedCategories = [];
     }
 }
