+++ project-base/app/src/Model/Product/Parameter/ParameterRepository.php
@@ -5,25 +5,25 @@ declare(strict_types=1);
 namespace App\Model\Product\Parameter;
 
 use App\Model\Product\Parameter\Exception\ParameterGroupNotFoundException;
-use App\Model\Product\Product;
+use App\Model\Product\Parameter\Exception\ParameterValueNotFoundException;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\Query\Expr\Join;
 use Doctrine\ORM\QueryBuilder;
-use Shopsys\FrameworkBundle\Component\Doctrine\OrderByCollationHelper;
-use Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig;
 use Shopsys\FrameworkBundle\Model\Category\Category;
 use Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterRepository as BaseParameterRepository;
 use Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValue;
 use Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValueData;
 use Shopsys\FrameworkBundle\Model\Product\Parameter\ProductParameterValue;
 use Shopsys\FrameworkBundle\Model\Product\Product as BaseProduct;
-use Shopsys\FrameworkBundle\Model\Product\ProductCategoryDomain;
 
 /**
  * @method \App\Model\Product\Parameter\Parameter|null findById(int $parameterId)
  * @method \App\Model\Product\Parameter\Parameter getById(int $parameterId)
+ * @method \App\Model\Product\Parameter\Parameter[] getParametersUsedByProductsInCategory(\App\Model\Category\Category $category, \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig)
+ * @method applyCategorySeoConditions(\Doctrine\ORM\QueryBuilder $queryBuilder, \App\Model\Category\Category $category, int $domainId)
  * @method \App\Model\Product\Parameter\Parameter getByUuid(string $uuid)
  * @method \App\Model\Product\Parameter\Parameter[] getAll()
+ * @method \App\Model\Product\Parameter\Parameter[] getAllWithTranslations(string $locale)
  * @method \Shopsys\FrameworkBundle\Model\Product\Parameter\ProductParameterValue[] getProductParameterValuesByProduct(\App\Model\Product\Product $product)
  * @method \Shopsys\FrameworkBundle\Model\Product\Parameter\ProductParameterValue[] getProductParameterValuesByProductSortedByOrderingPriorityAndName(\App\Model\Product\Product $product, string $locale)
  * @method string[][] getParameterValuesIndexedByProductIdAndParameterNameForProducts(\App\Model\Product\Product[] $products, string $locale)
@@ -31,38 +31,15 @@ use Shopsys\FrameworkBundle\Model\Product\ProductCategoryDomain;
  * @method \App\Model\Product\Parameter\Parameter|null findParameterByNames(string[] $namesByLocale)
  * @method \App\Model\Product\Parameter\Parameter[] getParametersByUuids(string[] $uuids)
  * @method \App\Model\Product\Parameter\Parameter[] getVisibleParametersByIds(int[] $parameterIds, string $locale)
- * @method \App\Model\Product\Product[] getFirstTenProductsWithSliderParameterValuesWithoutTheirsNumericValueFilled(string $locale)
  * @method \Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValue[] getParameterValuesByParameter(\App\Model\Product\Parameter\Parameter $parameter)
  * @method updateParameterValueInProductsByConversion(\App\Model\Product\Parameter\Parameter $parameter, \Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValue $oldParameterValue, \Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValue $newParameterValue)
  * @method \App\Model\Product\Parameter\Parameter[] getSliderParametersWithoutTheirsNumericValueFilled()
  * @method int getCountOfParameterValuesWithoutTheirsNumericValueFilledQueryBuilder(\App\Model\Product\Parameter\Parameter $parameter)
  * @method \App\Model\Product\Product[] getProductsByParameterValues(\Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValue[] $parameterValues)
- * @method \App\Model\Product\Parameter\Parameter[] getAllWithTranslations(string $locale)
  * @method bool existsParameterByName(string $name, string $locale, \App\Model\Product\Parameter\Parameter|null $excludeParameter = null)
  */
 class ParameterRepository extends BaseParameterRepository
 {
-    /**
-     * @param \App\Model\Category\Category $category
-     * @param \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig $domainConfig
-     * @return \App\Model\Product\Parameter\Parameter[]
-     */
-    public function getParametersUsedByProductsInCategory(Category $category, DomainConfig $domainConfig): array
-    {
-        $queryBuilder = $this->getParameterRepository()->createQueryBuilder('p')
-            ->select('p')
-            ->join(ProductParameterValue::class, 'ppv', Join::WITH, 'p = ppv.parameter')
-            ->join('p.translations', 'pt', Join::WITH, 'pt.locale = :locale')
-            ->setParameter('locale', $domainConfig->getLocale())
-            ->orderBy('p.orderingPriority', 'DESC')
-            ->addOrderBy(OrderByCollationHelper::createOrderByForLocale('pt.name', $domainConfig->getLocale()))
-            ->groupBy('p, pt');
-
-        $this->applyCategorySeoConditions($queryBuilder, $category, $domainConfig->getId());
-
-        return $queryBuilder->getQuery()->execute();
-    }
-
     /**
      * @param \App\Model\Category\Category $category
      * @param int $domainId
@@ -109,19 +86,20 @@ class ParameterRepository extends BaseParameterRepository
     }
 
     /**
-     * @param \Doctrine\ORM\QueryBuilder $queryBuilder
-     * @param \App\Model\Category\Category $category
-     * @param int $domainId
+     * @param int $parameterValueId
+     * @return \Shopsys\FrameworkBundle\Model\Product\Parameter\ParameterValue
      */
-    private function applyCategorySeoConditions(QueryBuilder $queryBuilder, Category $category, int $domainId): void
+    public function getParameterValueById(int $parameterValueId): ParameterValue
     {
-        $queryBuilder
-            ->join(Product::class, 'product', Join::WITH, 'ppv.product = product')
-            ->join(ProductCategoryDomain::class, 'pcd', Join::WITH, 'product = pcd.product')
-            ->andWhere('pcd.category = :category')
-            ->andWhere('pcd.domainId = :domainId')
-            ->setParameter('category', $category)
-            ->setParameter('domainId', $domainId);
+        $parameterValue = $this->getParameterValueRepository()->find($parameterValueId);
+
+        if ($parameterValue === null) {
+            $message = 'ParameterValue with ID ' . $parameterValueId . ' not found.';
+
+            throw new ParameterValueNotFoundException($message);
+        }
+
+        return $parameterValue;
     }
 
     /**
